requirements.txt
boto3==1.35.79
celery==5.4.0
channels==4.2.0
coverage==7.6.10
cryptography==44.0.0
daphne==4.1.2
Django==4.2.17
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-ipware==7.0.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
einops==0.8.0
gunicorn==23.0.0
numpy==1.26.4
openai==1.59.7
pandas==2.2.3
pdfplumber==0.11.5
pinecone-client==5.0.1
pillow==11.0.0
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pymupdf==1.25.3
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
python-multipart==0.0.19
pyyaml==6.0.2
replicate==1.0.7
openai==1.59.7
stripe==11.4.1
torch==2.8.0
transformers==4.55.0
uvicorn==0.34.0
websockets==14.2
django-countries==7.6.1
arrow==1.3.0
deep-translator==1.11.4
django-filter==25.1
openpyxl==3.1.5
scikit-learn==1.6.1
langdetect==1.0.9
sentence_transformers==4.1.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
